launch: factor out the YAML handling into new yaml_params module
